#969 refactor calendar component
