+++ __test__/components/EventModifications.test.tsx
@@ -1,7 +1,7 @@
 import * as appHooks from '@common/app/hooks'
 import * as EventDao from '@common/features/Events/EventDao'
 import { AppDispatch } from '@common/app/store'
-import CalendarApp from '@common/components/Calendar/Calendar'
+import CalendarLayout from '@private/components/Calendar/CalendarLayout'
 import {
   createEventHandlers,
   EventHandlersProps
@@ -80,17 +80,7 @@ describe('CalendarApp integration', () => {
     }
 
     const mockCalendarRef = { current: null }
-    renderWithProviders(
-      <CalendarApp
-        setCurrentView={jest.fn()}
-        onViewChange={jest.fn()}
-        openSidebar={false}
-        onCloseSidebar={jest.fn()}
-        currentView="timeGridWeek"
-        calendarRef={mockCalendarRef}
-      />,
-      preloadedState
-    )
+    renderWithProviders(<CalendarLayout />, preloadedState)
   }
 
   it('renders the event on the calendar and calendarRef works', async () => {
@@ -186,17 +176,7 @@ describe('CalendarApp integration', () => {
       title: 'Private Event'
     })
     const mockCalendarRef = { current: null }
-    renderWithProviders(
-      <CalendarApp
-        setCurrentView={jest.fn()}
-        onViewChange={jest.fn()}
-        openSidebar={false}
-        onCloseSidebar={jest.fn()}
-        currentView="timeGridWeek"
-        calendarRef={mockCalendarRef}
-      />,
-      preloadedState
-    )
+    renderWithProviders(<CalendarLayout />, preloadedState)
     const card = screen.getByTestId('event-card-event1')
     const lockIcon = within(card).getByTestId('LockOutlinedIcon')
     expect(lockIcon).toBeInTheDocument()
@@ -212,17 +192,7 @@ describe('CalendarApp integration', () => {
       title: 'Confidential Event'
     })
     const mockCalendarRef = { current: null }
-    renderWithProviders(
-      <CalendarApp
-        setCurrentView={jest.fn()}
-        onViewChange={jest.fn()}
-        openSidebar={false}
-        onCloseSidebar={jest.fn()}
-        currentView="timeGridWeek"
-        calendarRef={mockCalendarRef}
-      />,
-      preloadedState
-    )
+    renderWithProviders(<CalendarLayout />, preloadedState)
 
     const card = screen.getByTestId('event-card-event1')
     const lockIcon = within(card).getByTestId('LockOutlinedIcon')
@@ -239,17 +209,7 @@ describe('CalendarApp integration', () => {
       title: 'Public Event'
     })
     const mockCalendarRef = { current: null }
-    renderWithProviders(
-      <CalendarApp
-        setCurrentView={jest.fn()}
-        onViewChange={jest.fn()}
-        openSidebar={false}
-        onCloseSidebar={jest.fn()}
-        currentView="timeGridWeek"
-        calendarRef={mockCalendarRef}
-      />,
-      preloadedState
-    )
+    renderWithProviders(<CalendarLayout />, preloadedState)
 
     const card = screen.getByTestId('event-card-event1')
     const lockIcon = within(card).queryByTestId('LockOutlinedIcon')
@@ -258,56 +218,46 @@ describe('CalendarApp integration', () => {
 
   it('does render a title for events without any attendees or user as organizer', async () => {
     const mockCalendarRef = { current: null }
-    renderWithProviders(
-      <CalendarApp
-        setCurrentView={jest.fn()}
-        onViewChange={jest.fn()}
-        openSidebar={false}
-        onCloseSidebar={jest.fn()}
-        currentView="timeGridWeek"
-        calendarRef={mockCalendarRef}
-      />,
-      {
-        user: {
-          userData: {
-            sub: 'test',
-            email: 'test@test.com',
-            sid: 'mockSid',
-            openpaasId: '667037022b752d0026472254'
-          },
-          tokens: {
-            accessToken: 'token'
-          }
+    renderWithProviders(<CalendarLayout />, {
+      user: {
+        userData: {
+          sub: 'test',
+          email: 'test@test.com',
+          sid: 'mockSid',
+          openpaasId: '667037022b752d0026472254'
         },
-        calendars: {
-          list: {
-            '667037022b752d0026472254/cal1': {
-              name: 'Calendar 1',
-              id: '667037022b752d0026472254/cal1',
-              color: { light: '#FF0000', dark: '#000' },
-              owner: { emails: ['alice@example.com'] },
-              events: {
-                event1: {
-                  id: 'event1',
-                  calId: '667037022b752d0026472254/cal1',
-                  uid: 'event1',
-                  start: new Date().toISOString(),
-                  end: new Date(Date.now() + 3600000).toISOString(),
-                  partstat: 'ACCEPTED',
-                  organizer: {
-                    cn: 'Alice',
-                    cal_address: 'alice@example.com'
-                  },
-                  class: 'PUBLIC',
-                  title: 'Public Event'
-                }
+        tokens: {
+          accessToken: 'token'
+        }
+      },
+      calendars: {
+        list: {
+          '667037022b752d0026472254/cal1': {
+            name: 'Calendar 1',
+            id: '667037022b752d0026472254/cal1',
+            color: { light: '#FF0000', dark: '#000' },
+            owner: { emails: ['alice@example.com'] },
+            events: {
+              event1: {
+                id: 'event1',
+                calId: '667037022b752d0026472254/cal1',
+                uid: 'event1',
+                start: new Date().toISOString(),
+                end: new Date(Date.now() + 3600000).toISOString(),
+                partstat: 'ACCEPTED',
+                organizer: {
+                  cn: 'Alice',
+                  cal_address: 'alice@example.com'
+                },
+                class: 'PUBLIC',
+                title: 'Public Event'
               }
             }
-          },
-          pending: false
-        }
+          }
+        },
+        pending: false
       }
-    )
+    })
 
     expect(screen.getByText('Public Event')).toBeInTheDocument()
   })
