+++ common/src/app/selectors/selectCalendars.tsx
@@ -3,5 +3,5 @@ import { createSelector } from '@reduxjs/toolkit'
 
 export const selectCalendars = createSelector(
   (state: RootState) => state.calendars.list,
-  list => Object.values(list)
+  list => Object.values(list || {})
 )
