+++ common/src/window.d.ts
@@ -40,7 +40,7 @@ declare global {
 
     appList: AppIconProps[]
 
-    __calendarRef: MutableRefObject<CalendarApi | null>
+    __calendarRef?: MutableRefObject<CalendarApi | null>
 
     DISABLE_PUBLIC_VISIBILITY: boolean
